requirements.txt
streamlit>=1.28.0
python-pptx>=0.6.19
python-dotenv>=0.21.1
google-generativeai>=0.3.0
Pillow>=9.5.0
requests>=2.28.2
comtypes>=1.1.14
pywin32>=306